misc: move AgentRunContext to Response folder
