+++ sdk/agentserver/Azure.AI.AgentServer.Core/src/Responses/Endpoint/AgentInvoker.cs
@@ -5,7 +5,6 @@ using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using Azure.AI.AgentServer.Contracts.Generated.OpenAI;
 using Azure.AI.AgentServer.Contracts.Generated.Responses;
-using Azure.AI.AgentServer.Core.AgentRun;
 using Azure.AI.AgentServer.Core.Common.Http.ServerSentEvent;
 using Azure.AI.AgentServer.Core.Telemetry;
 using Microsoft.AspNetCore.Http;
