+++ sdk/agentserver/Azure.AI.AgentServer.Core/src/Responses/Endpoint/AgentRunEndpoints.cs
@@ -3,7 +3,6 @@
 
 using Azure.AI.AgentServer.Contracts.Generated.OpenAI;
 using Azure.AI.AgentServer.Contracts.Generated.Responses;
-using Azure.AI.AgentServer.Core.AgentRun;
 using Azure.AI.AgentServer.Core.Telemetry;
 using Azure.AI.AgentServer.Responses.Invocation;
 using Microsoft.AspNetCore.Builder;
