+++ sdk/agentserver/Azure.AI.AgentServer.Core/src/Responses/Invocation/AgentInvocationContext.cs
@@ -12,11 +12,11 @@ namespace Azure.AI.AgentServer.Responses.Invocation;
 /// <param name="responseId">The response ID for this invocation.</param>
 /// <param name="conversationId">The conversation ID for this invocation.</param>
 /// <remarks>
-/// NOTE: This class is deprecated. Use <see cref="Core.AgentRun.AgentRunContext"/> instead,
+/// NOTE: This class is deprecated. Use <see cref="AgentRunContext"/> instead,
 /// which provides richer context including request data, user information, and tool definitions.
 /// See the MIGRATION_GUIDE.md for detailed migration instructions.
 /// </remarks>
-[Obsolete("AgentInvocationContext is deprecated. Use AgentRunContext from Azure.AI.AgentServer.Core.AgentRun namespace instead. See MIGRATION_GUIDE.md for details.", false)]
+[Obsolete("AgentInvocationContext is deprecated. Use AgentRunContext from Azure.AI.AgentServer.Responses.Invocation namespace instead. See MIGRATION_GUIDE.md for details.", false)]
 public class AgentInvocationContext(IIdGenerator idGenerator,
     string responseId,
     string conversationId)
