+++ sdk/agentserver/Azure.AI.AgentServer.Core/src/Responses/Invocation/ResponsesExtensions.cs
@@ -5,7 +5,6 @@ using Azure.AI.AgentServer.Contracts.Generated.Agents;
 using Azure.AI.AgentServer.Contracts.Generated.Conversations;
 using Azure.AI.AgentServer.Contracts.Generated.OpenAI;
 using Azure.AI.AgentServer.Contracts.Generated.Responses;
-using Azure.AI.AgentServer.Core.AgentRun;
 using Azure.AI.AgentServer.Core.Common.Http.Json;
 
 namespace Azure.AI.AgentServer.Responses.Invocation;
