+++ sdk/agentserver/Azure.AI.AgentServer.Core/src/Responses/Invocation/Stream/NestedResponseGenerator.cs
@@ -3,7 +3,6 @@
 
 using Azure.AI.AgentServer.Contracts.Generated.OpenAI;
 using Azure.AI.AgentServer.Contracts.Generated.Responses;
-using Azure.AI.AgentServer.Core.AgentRun;
 
 namespace Azure.AI.AgentServer.Responses.Invocation.Stream;
