+++ sdk/agentserver/Azure.AI.AgentServer.Core/tests/Agents.Customized.Integration.Tests/CustomizedAgentInvocation.cs
@@ -1,7 +1,6 @@
 using System.Runtime.CompilerServices;
 using Azure.AI.AgentServer.Contracts.Generated.OpenAI;
 using Azure.AI.AgentServer.Contracts.Generated.Responses;
-using Azure.AI.AgentServer.Core.AgentRun;
 using Azure.AI.AgentServer.Core.Common.Http.Json;
 using Azure.AI.AgentServer.Core.Common.Id;
 using Azure.AI.AgentServer.Responses.Invocation;
